Fix SSL connection issues.
There were 2 issues causing the incorrect SSL connection.

The 1st issue was the `openssl s_client` was in the waiting status in the
script. As a solution, redirected the command's stdin to `/dev/null`. The
`openssl s_client` waited for input when it connected successfully. This also
fixes the issue that you need to type enter a few times to exit the command.

The 2nd issue was that the `SECLEVEL=1` defined in the file below prevents a SSL
client from connecting with SHA1-MD5 digest and TLS < 1.2.

/usr/share/crypto-policies/<secuirty_policy>/opensslcnf.txt

Updated the log files and documents by this change too.